Move object detection into its own module
