+++ src/detection/rv/od/commands/make_predict_chips.py
@@ -6,8 +6,8 @@ import click
 import numpy as np
 import rasterio
 
-from rv.commands.utils import load_window, save_img
-from rv.commands.settings import planet_channel_order
+from rv.od.commands.utils import load_window, save_img
+from rv.od.commands.settings import planet_channel_order
 
 
 def _make_predict_chips(image_path, chips_dir, chips_info_path,
