+++ src/detection/rv/od/commands/predict.py
@@ -3,12 +3,12 @@ from shutil import move
 
 import click
 
-from rv.commands.make_predict_chips import _make_predict_chips
-from rv.commands.predict_on_chips import _predict_on_chips
-from rv.commands.aggregate_predictions import _aggregate_predictions
-from rv.commands.transform_geojson import _transform_geojson
-from rv.commands.settings import planet_channel_order, temp_root_dir
-from rv.commands.utils import (
+from rv.od.commands.make_predict_chips import _make_predict_chips
+from rv.od.commands.predict_on_chips import _predict_on_chips
+from rv.od.commands.aggregate_predictions import _aggregate_predictions
+from rv.od.commands.transform_geojson import _transform_geojson
+from rv.od.commands.settings import planet_channel_order, temp_root_dir
+from rv.od.commands.utils import (
     download_if_needed, upload_if_needed, get_local_path, make_temp_dir,
     download_and_build_vrt)
